% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{Balance}
\alias{Balance}
\title{Balancing of classes in multi-modal modelling}
\description{
Provides a pre-processing wrapper to balance classes during modelling using the Synthetic Minority Oversampling Technique (SMOTE).
}
\details{
To implement SMOTE for a multi-class problem, balancing is performed on each class.
}
