% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{cpoImputeData}
\alias{cpoImputeData}
\title{Impute data using a CPO}
\usage{
cpoImputeData(
  impute_method = "MICE",
  id,
  export = "export.default",
  affect.type = NULL,
  affect.index = integer(0),
  affect.names = character(0),
  affect.pattern = NULL,
  affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE
)
}
\arguments{
\item{impute_method}{(character)\cr
Method of imputation - \'MICE\' or \'KNN\'}
}
\value{
A pre-processing object that  can be used in a pipeline to perform imputation.
}
\description{
Create a pre-processing object to perform imputation in the ML pipeline.
Two imputation methods are provided: Mice and KNN.
}
\examples{
lrn <- mlr::makeLearner(cl = "classif.gbm", id = "test", predict.type = "prob")
lrn <- cpoImputeData("MICE") \%>>\% lrn
}
