% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{cpoNormalise}
\alias{cpoNormalise}
\title{Normalise data using a CPO}
\usage{
cpoNormalise(
  method = "STAND",
  id,
  export = "export.default",
  affect.type = NULL,
  affect.index = integer(0),
  affect.names = character(0),
  affect.pattern = NULL,
  affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE
)
}
\arguments{
\item{method}{(character)\cr
Method of normalisation}
}
\value{
Nothing but the function can be used in a pipeline to perform normalisation.
}
\description{
Create a pre-processing object to perform normalisation in the ML pipeline.
Eight different methods of normalisation are available: 
    MINMAX, LOGT, STAND, ZSCORE, MEDIAN, CPM, CPM_LOGT, QUANTILE
Normalisation is only applied to the numeric columns (non-boolean)
}
\examples{
lrn <- mlr::makeLearner(cl = "classif.gbm", id = "test", predict.type = "prob")
lrn <- cpoNormalise("ZSCORE") \%>>\% lrn
}
