% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Features}
\alias{Features}
\title{R6 class to hold feature selections}
\description{
Stores feature selections and gives access to feature importance scores.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{featsel}}{(list)\cr
Record of the features selected by each model in each fold of cross-validation.
Each element of the list is a data.frame, containing the raw feature selections for one modality.
The data.frame contains one row per feature and one column per cross-validation fold.}

\item{\code{featsel_aggr}}{(list)\cr
Record of features selected, aggregated across all folds of cross-validation.
Each element of the list is a data.frame, containing the aggregated feature selections for one modality.
The data.frame contains one row per feature and a single column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Features-new}{\code{Features$new()}}
\item \href{#method-Features-getFeatImpScores}{\code{Features$getFeatImpScores()}}
\item \href{#method-Features-save}{\code{Features$save()}}
\item \href{#method-Features-save_multiclass}{\code{Features$save_multiclass()}}
\item \href{#method-Features-complete}{\code{Features$complete()}}
\item \href{#method-Features-get_results}{\code{Features$get_results()}}
\item \href{#method-Features-write}{\code{Features$write()}}
\item \href{#method-Features-clone}{\code{Features$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-new"></a>}}
\if{latex}{\out{\hypertarget{method-Features-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Features` object.
Extracts the features names from the tasks for each modality.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$new(tasks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tasks}}{(list)\cr
List of tasks (datasets) for the multi-modal model - one per modality}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Features` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-getFeatImpScores"></a>}}
\if{latex}{\out{\hypertarget{method-Features-getFeatImpScores}{}}}
\subsection{Method \code{getFeatImpScores()}}{
Extract the feature importance scores from an mlr model.
Convert the custom importance data from each model into a uniform format -
a data.frame with one row per feature and one column per class, 
plus a column labelled 'all', which gives the feature importance for the model as a whole.
The 'all' column may contain zeroes if this data is not available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$getFeatImpScores(mod, class_names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{(???)\cr
The mlr model}

\item{\code{class_names}}{(character vector)
Names of the classes in the data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame containing feature importance scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-save"></a>}}
\if{latex}{\out{\hypertarget{method-Features-save}{}}}
\subsection{Method \code{save()}}{
Retrieve and save the feature importance scores from one mlr model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$save(mod, task, classes, method, fold_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{(???)\cr
The mlr model}

\item{\code{task}}{(Task)\cr
The task for which the scores were calculated.}

\item{\code{classes}}{(character vector)
Names of the classes in the data}

\item{\code{method}}{(character)\cr
The name of the method used ???}

\item{\code{fold_num}}{(integer)\cr
The index of the fold for which the scores were calculated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame containing feature importance scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-save_multiclass"></a>}}
\if{latex}{\out{\hypertarget{method-Features-save_multiclass}{}}}
\subsection{Method \code{save_multiclass()}}{
Save a set of raw feature importance scores from a single multi-class task for one cross-validation fold.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$save_multiclass(method, results, task_id, fold_num, seln_col = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{(character)\cr
The name of the method used ???}

\item{\code{results}}{(data.frame)\cr
A data.frame containing feature importance score, one column per class, one row per feature.}

\item{\code{task_id}}{(character)\cr
The name of the task for which the scores were calculated.}

\item{\code{fold_num}}{(integer)\cr
The index of the fold for which the scores were calculated.}

\item{\code{seln_col}}{(character)\cr
The name of the column in the results data.frame containing the scores for the model as a whole}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame containing feature importance scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-complete"></a>}}
\if{latex}{\out{\hypertarget{method-Features-complete}{}}}
\subsection{Method \code{complete()}}{
The scores are  normalised across all modalities so that they can be compared.
Only features apearing in at least 75% of cross-vaidated runs are given a score.
The scores are also aggregated by averaging across all cross-validation runs. 
The scores from the individual cross-validation runs are saved to a file ending in '_featsel.csv'.
The aggregated scores are saved to a file ending in '_featsel_aggr.csv'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$complete()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-Features-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the selected features - for each fold and an aggregated result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$get_results(perc = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{perc}}{(numeric)\cr
   Only features which are selected in this percentage of folds or greater are returned.
To retrieve all selected features, pass 1??}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-write"></a>}}
\if{latex}{\out{\hypertarget{method-Features-write}{}}}
\subsection{Method \code{write()}}{
Write the selected features to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$write(result_file, suffix = "", perc = 0.75)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result_file}}{(character)\cr
The prefix of the name of the files to which the results will be written.}

\item{\code{suffix}}{(character)\cr
The suffix will be appended to the prefix of the output file name.}

\item{\code{perc}}{(numeric)\cr
   Only features which are selected in this percentage of folds or greater are returned.
To retrieve all selected features, pass 1??}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Features-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Features-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Features$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
