% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{Filters}
\alias{Filters}
\title{Filters: R6 class representing the custom mlr filters used by the models}
\description{
Register the custom filters with mlr
}
\details{
Three custom filters are provided. 
Filter boruta uses the Boruta method to select important features.
Filter randomForestSRC uses randomForestSRC to select important features.
Filter \dQuote{randomForestSRC_var.select} uses the minimal depth variable
selection proposed by Ishwaran et al. (2010) (`method = "md"`) or a
variable hunting approach (`method = "vh"` or `method = "vh.vimp"`).
The minimal depth measure is the default.
The randomForestSRC and randomForestSRC_var.select were deleted from mlr.
The filters are registered so they can be used 
with makeFilterWrapper and other mlr filter functions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Filters-new}{\code{Filters$new()}}
\item \href{#method-Filters-getImpSHAP}{\code{Filters$getImpSHAP()}}
\item \href{#method-Filters-getImpRFSRC}{\code{Filters$getImpRFSRC()}}
\item \href{#method-Filters-clone}{\code{Filters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filters-new"></a>}}
\if{latex}{\out{\hypertarget{method-Filters-new}{}}}
\subsection{Method \code{new()}}{
Create the custom filters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filters$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing but the filters will be registered
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filters-getImpSHAP"></a>}}
\if{latex}{\out{\hypertarget{method-Filters-getImpSHAP}{}}}
\subsection{Method \code{getImpSHAP()}}{
Method to obtain feature inportance scores using Shapley values.
For use with the boruta custom filter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filters$getImpSHAP(x, y, task, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Data frame of predictors including shadows}

\item{\code{y}}{The response vector}

\item{\code{task}}{The mlr task}

\item{\code{...}}{Parameters passed to the underlying call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Feature importance scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filters-getImpRFSRC"></a>}}
\if{latex}{\out{\hypertarget{method-Filters-getImpRFSRC}{}}}
\subsection{Method \code{getImpRFSRC()}}{
Method to obtain feature inportance scores using randomForestSRC.
For use with the boruta custom filter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filters$getImpRFSRC(x, y, ntree = 500, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Data frame of predictors including shadows}

\item{\code{y}}{The response vector}

\item{\code{ntree}}{The number of trees to use in the random forest.}

\item{\code{...}}{Parameters passed to the underlying call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Feature importance scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Filters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
