% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{Imputation}
\alias{Imputation}
\title{Imputation methods for use with multi-modal modelling}
\description{
Provides two imputation methods for adding imputation to the pre-processing pipeline: Mice and KNN.
}
\details{
Two methods are provided for adding imputation to the machine learning pipeline.
One uses mlrCPO to create a new Composable Preprocessing Operator (CPO) for imputation.
The other creates a preprocessing wrapper that can be added to a learner.
}
