% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipoproteins}
\alias{lipoproteins}
\title{Lipoproteins}
\format{
A data frame with 106 rows (samples) and 48 columns (features)
}
\usage{
data(lipoproteins)
}
\description{
The Lipoprotein data modality.
}
\keyword{datasets}
