% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{make_config}
\alias{make_config}
\title{Create a configuration object}
\usage{
make_config(filename)
}
\arguments{
\item{filename}{(character)\cr
The name of teh configuration file.}
}
\value{
A configuration object.
}
\description{
Reads in the supplied configuration file, which contains information about how the model should be constructed,
and sets up a configuration object for use by the models.
}
