% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{makePreprocWrapperBalanceMC}
\alias{makePreprocWrapperBalanceMC}
\title{Balance data using a pre-processing wrapper}
\usage{
makePreprocWrapperBalanceMC(learner, target_name, bal_method)
}
\arguments{
\item{learner}{(character)\cr
The learner to which balancing should be added.}

\item{target_name}{(character)\cr
The name of teh target variable in the data.}

\item{bal_method}{(character)\cr
Method of balancing - \'SMOTE\' or \'OVERSAMPLE\'}
}
\value{
A pre-processing wrapper. The function can be used in a pipeline to perform balancing.
}
\description{
Create a pre-processing wrapper to balance classes in the ML pipeline using the Synthetic Minority Oversampling Technique (SMOTE).
To implement SMOTE for a multi-class problem, balancing is performed on each class.
}
\examples{
lrn <- mlr::makeLearner(cl = "classif.gbm", id = "test", predict.type = "prob")
lrn <- makePreprocWrapperBalanceMC(lrn, "ID", "SMOTE")
}
