% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{makePreprocWrapperImpute}
\alias{makePreprocWrapperImpute}
\title{Impute data using a pre-processing wrapper}
\usage{
makePreprocWrapperImpute(learner, impute_method = "MICE")
}
\arguments{
\item{learner}{(character)\cr
The learner to which imputation should be added.}

\item{impute_method}{(character)\cr
Method of imputation - \'MICE\' or \'KNN\'}
}
\value{
A pre-processing wrapper. The function can be used in a pipeline to perform imputation.
}
\description{
Create a pre-processing wrapper to perform imputation in the ML pipeline.
Two imputation methods are provided: Mice and KNN.
}
\examples{
lrn <- mlr::makeLearner(cl = "classif.gbm", id = "test", predict.type = "prob")
lrn <- makePreprocWrapperImpute(lrn, "KNN")
}
