% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{makePreprocWrapperNormalise}
\alias{makePreprocWrapperNormalise}
\title{Normalise data using a pre-processing wrapper}
\usage{
makePreprocWrapperNormalise(learner, method = "STAND")
}
\arguments{
\item{learner}{(character)\cr
The learner to which imputation should be added.}

\item{method}{(character)\cr
Method of normalisation}
}
\value{
A pre-processing wrapper. The function can be used in a pipeline to perform normalisation.
}
\description{
Create a pre-processing wrapper to perform normalisation in the ML pipeline.
Eight different methods of normalisation are available: 
    MINMAX, LOGT, STAND, ZSCORE, MEDIAN, CPM, CPM_LOGT, QUANTILE#' Normalisation is only applied to the numeric columns (non-boolean)
}
\examples{
lrn <- mlr::makeLearner(cl = "classif.gbm", id = "test", predict.type = "prob")
lrn <- makePreprocWrapperImpute(lrn, "MINMAX")
}
