% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_adaboost.R
\docType{package}
\name{MM_Adaboost}
\alias{MM_Adaboost}
\title{R6 Class representing a multi-modal Adaboost classifier}
\description{
Creates a multi-modal Adaboost classifier that uses hard or 
soft voting or a meta learner to determine the final outcome of the classifier.
}
\details{
For each step in the boosting process, this model
trains a classifier on each modality then combines the
predictions from those modalities using a hard vote
(majority voting), soft vote (average of probabilities) 
or a meta learner to give the final prediction.
After each boosting step, the weights of the samples are updated
to ensure that the samples that were misclassified in the last step
are given more weight in subsequent boosting steps.
}
