% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{MM_Config}
\alias{MM_Config}
\title{MM_Config: R6 class representing the model configuration}
\description{
Provides functions to read and store configuration parameters.
The configuration specifies how the model should be constructed.
}
