% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_incremental.R
\docType{package}
\name{MM_Incremental}
\alias{MM_Incremental}
\title{R6 Class representing an incremental model to determine
an optimal subset of modalities for model building.}
\description{
Forward/backward sequential building of a sequence of views.
Only works for classification.
}
