% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_meta_learner.R
\name{MM_Meta_Learner}
\alias{MM_Meta_Learner}
\title{R6 Class representing a multi-modal meta learner classifier}
\description{
Creates a multi-modal meta learner classifier that uses
soft voting to determine the final outcome of the classifier.
}
\details{
Still to come
}
