% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_model.R
\name{MM_Model}
\alias{MM_Model}
\title{MM_Model: R6 class representing a multi-modal model}
\description{
Provides functions to create and fit a generic model, which is run using mlr.
This is a super-class of all of the multi-model models.
}
