% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_results.R
\name{MM_Results}
\alias{MM_Results}
\title{R6 Class to store the results for the multi-modal models}
\description{
A container for the different types of results returned by the multi-modal models - \cr
performance, roc measures, selected features, feature stability and predictions.
}
