% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_single.R
\docType{package}
\name{MM_Single}
\alias{MM_Single}
\title{R6 Class representing a classifier or survival analysis model for single modalities}
\description{
Creates a classifier or or survival analysis model
for each individual modality.
}
\details{
Trains a classifier or survival analysis model on each modality.
}
