% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_voting.R
\name{MM_Voting}
\alias{MM_Voting}
\title{MM_Voting: R6 Class representing a multi-modal voting ensemble classifier}
\description{
Creates a multi-modal ensemble classifier that uses 
hard or soft voting to determine the final outcome of the classifier.
}
\details{
Trains a classifier on each modality then combines the
predictions from those modalities using a hard vote
(majority voting) or a soft vote (average of probabilities)
to give a final prediction.
}
