% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{Normalisation}
\alias{Normalisation}
\title{Normalisation methods for use with multi-modal modelling}
\description{
Provides methods for adding normalisation to the pre-processing pipeline.
}
\details{
Two methods are provided for adding normalisation to the machine learning pipeline.
One uses mlrCPO to create a new Composable Preprocessing Operator (CPO) for normalisation.
The other creates a preprocessing wrapper that can be added to a learner.
Normalisation is only applied to the numeric columns (non-boolean)
}
