% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OralSpecies}
\alias{OralSpecies}
\title{OralSpecies}
\format{
A data frame with 106 rows (samples) and 48 columns (features)
}
\description{
The Oral Microbiome data modality - species only.
}
\keyword{data}
