% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathology}
\alias{pathology}
\title{Pathology}
\format{
A data frame with 106 rows (samples) and 48 columns (features)
}
\usage{
data(pathology)
}
\description{
The Pathology data modality.
}
\keyword{datasets}
