% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{Performance}
\alias{Performance}
\title{R6 class to hold the performance results}
\description{
Stores and gives access to the performance results for each metric for each modality.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{perf}}{(list of lists)\cr
The top level list contains one list per performance metric requested.
The second level list contains one value per modality ???}

\item{\code{aggr}}{(list of lists)\cr
The top level list contains one list per performance metric requested.
The second level list contains teh aggregated performance for each modality.}

\item{\code{measures}}{(list)\cr
The requested mlr performance metrics.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Performance-new}{\code{Performance$new()}}
\item \href{#method-Performance-calculate}{\code{Performance$calculate()}}
\item \href{#method-Performance-save}{\code{Performance$save()}}
\item \href{#method-Performance-write}{\code{Performance$write()}}
\item \href{#method-Performance-get_results}{\code{Performance$get_results()}}
\item \href{#method-Performance-plot}{\code{Performance$plot()}}
\item \href{#method-Performance-clone}{\code{Performance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Performance-new"></a>}}
\if{latex}{\out{\hypertarget{method-Performance-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Performance` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Performance$new(measures)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measures}}{(list)\cr
The names of the requested mlr performance metrics.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Performance` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Performance-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-Performance-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate the performance from an mlr Prediction object and save it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Performance$calculate(pred, task = NULL, model = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pred}}{(Prediction)\cr
mlr Prediction object}

\item{\code{task}}{(Task)\cr
mlr Task object}

\item{\code{model}}{(object)\cr
mlr trained model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Performance-save"></a>}}
\if{latex}{\out{\hypertarget{method-Performance-save}{}}}
\subsection{Method \code{save()}}{
Save the performance results internally
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Performance$save(res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{res}}{(ResampleResult or ???)\cr
An mlr structure containing the performance results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Performance-write"></a>}}
\if{latex}{\out{\hypertarget{method-Performance-write}{}}}
\subsection{Method \code{write()}}{
Write the performance results to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Performance$write(result_file, suffix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result_file}}{(character)\cr
The prefix of the name of the files to which the results will be written.}

\item{\code{suffix}}{(character)\cr
The suffix will be appended to the prefix of the output file name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Performance-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-Performance-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the performance results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Performance$get_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame containing the performance results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Performance-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Performance-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the performance results as boxplots and return the JPEG plot object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Performance$plot(method, result_file = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{(character)\cr
Method used ???}

\item{\code{result_file}}{(character)\cr
   The name of the file containing the results to be plotted (optional).
If this value is NA (the default), the results saved internally are plotted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The plot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Performance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Performance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Performance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
