% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_measures.R
\name{PerformanceMeasures}
\alias{PerformanceMeasures}
\title{PerformanceMeasures: R6 class representing custom performance measures}
\description{
Modifies existing performance measures to create new measures that provide a valid result in the presence of NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{measures}}{(list)
List of the performance measures to be used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PerformanceMeasures-new}{\code{PerformanceMeasures$new()}}
\item \href{#method-PerformanceMeasures-clone}{\code{PerformanceMeasures$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerformanceMeasures-new"></a>}}
\if{latex}{\out{\hypertarget{method-PerformanceMeasures-new}{}}}
\subsection{Method \code{new()}}{
Create a new PerformanceMeasures object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerformanceMeasures$new(task_type, predict_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_type}}{(character)\cr
Type of model - "classif" for classification, "multilabel" for multilabel classification or "surv" for survival analysis.}

\item{\code{predict_type}}{(character)\cr
The type of prediction to be made - “response” or “prob.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PerformanceMeasures` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerformanceMeasures-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PerformanceMeasures-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerformanceMeasures$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
