% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_measures.R
\docType{package}
\name{ROCMultiClass}
\alias{ROCMultiClass}
\title{R6 class to hold multi-class Receiver Operating Curve (ROC) measures}
\description{
Calculates and stores ROC measures for a multiclass classification problem.
}
