% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SmallMolecules}
\alias{SmallMolecules}
\title{SmallMolecules}
\format{
A data frame with 106 rows (samples) and 17 columns (features)
}
\description{
The SmallMolecules data modality.
}
\keyword{data}
