% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{Stability}
\alias{Stability}
\title{R6 class to hold the stability results}
\description{
Calculates and stores stability results for each stability metric.
}
\details{
Five stability metrics are available: the Jaccard Index, the Dice Score,
Kuncheva's Index, Lustgarten's Index and the Relative Weighted Consistency Index.
}
\references{
Relative weighted consistency index - Song et al 2019, Petr Somol and Jana Novovicova 2010
}
