% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StoolSpecies}
\alias{StoolSpecies}
\title{StoolSpecies}
\format{
A data frame with 106 rows (samples) and 48 columns (features)
}
\description{
The Stool Microbiome data modality - species only.
}
\keyword{data}
