% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R, R/imputation.R
\name{Imputation}
\alias{Imputation}
\title{Balancing of classes in multi-modal modelling}
\description{
Provides a pre-processing wrapper to balance classes during modelling using the Synthetic Minority Oversampling Technique (SMOTE).

Provides two imputation methods for adding imputation to the pre-processing pipeline: Mice and KNN.
}
\details{
To implement SMOTE for a multi-class problem, balancing is performed on each class.

Two methods are provided for adding imputation to the machine learning pipeline.
One uses mlrCPO to create a new Composable Preprocessing Operator (CPO) for imputation.
The other creates a preprocessing wrapper that can be added to a learner.
}
